# GestionStock/src/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Gestion de stock REST API
  version: "1.0"
servers:
- url: http://localhost:8080
  description: Generated server url
security:
- JavaInUseSecurityScheme: []
paths:
...
  /utilisateurs:
    get:
      tags:
      - Utilisateur
      operationId: findAllUtilisateurs
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageResponseUtilisateurResponse"
    post:
      tags:
      - Utilisateur
      operationId: saveUtilisateur
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UtilisateurRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: integer
                format: int64
  /utilisateurs/photo/upload/{id}:
    post:
      tags:
      - Utilisateur
      operationId: uploadUtilisateurPhoto
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
              - file
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
...
  /utilisateurs/mot-de-passe/update:
    patch:
      tags:
      - Utilisateur
      operationId: changeUtilisateurPassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: integer
                format: int64
...
  /utilisateurs/login/{login}:
    get:
      tags:
      - Utilisateur
      operationId: findUtilisateurByLogin
      parameters:
      - name: login
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UtilisateurResponse"
  /utilisateurs/id/{id}:
    get:
      tags:
      - Utilisateur
      operationId: findUtilisateurById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UtilisateurResponse"
...
components:
  schemas:
...
    AdresseRequest:
      type: object
      properties:
        adresse1:
          type: string
        adresse2:
          type: string
        ville:
          type: string
        codePostale:
          type: string
        pays:
          type: string
      required:
      - adresse1
      - pays
      - ville
    UtilisateurRequest:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nom:
          type: string
        sexe:
          type: string
          enum:
          - MASCULIN
          - FEMININ
        login:
          type: string
        numeroTel:
          type: string
        email:
          type: string
        adresse:
          $ref: "#/components/schemas/AdresseRequest"
      required:
      - adresse
      - login
      - nom
      - numeroTel
      - sexe
...
    AdresseResponse:
      type: object
      properties:
        adresse1:
          type: string
        adresse2:
          type: string
        ville:
          type: string
        codePostale:
          type: string
        pays:
          type: string
...
    UtilisateurResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nom:
          type: string
        login:
          type: string
        sexe:
          type: string
          enum:
          - MASCULIN
          - FEMININ
        numeroTel:
          type: string
        email:
          type: string
        photo:
          type: string
          format: byte
        adresse:
          $ref: "#/components/schemas/AdresseResponse"
...
    ChangePasswordRequest:
      type: object
      properties:
        utilisateur:
          type: integer
          format: int64
          minimum: 1
        motDePasse:
          type: string
        confirmMotDePasse:
          type: string
      required:
      - confirmMotDePasse
      - motDePasse
      - utilisateur
...
    PageResponseUtilisateurResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/UtilisateurResponse"
        number:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
...
  securitySchemes:
    JavaInUseSecurityScheme:
      type: http
      name: JavaInUseSecurityScheme
      scheme: basic